refactor: Update router configuration for authentication and protected routes
