+++ phone-test/src/router.tsx
@@ -1,0 +1,23 @@
+import { createBrowserRouter, createRoutesFromElements, Route } from 'react-router-dom';
+import { LoginPage } from './pages/Login/Login';
+import { CallsListPage } from './pages/CallsList';
+import { CallDetailsPage } from './pages/CallDetails';
+
+import { ProtectedLayout } from './components/routing/ProtectedLayout';
+import { AuthProvider } from './hooks/useAuth';
+import { AppRedirect } from './AppRedirect';
+
+const router = createBrowserRouter(
+  createRoutesFromElements(
+    <Route element={<AuthProvider />}>
+      <Route path="/" element={<AppRedirect />} />
+      <Route path="/login" element={<LoginPage />} />
+      <Route path="/calls" element={<ProtectedLayout />}>
+        <Route path="/calls" element={<CallsListPage />} />
+        <Route path="/calls/:callId" element={<CallDetailsPage />} />
+      </Route>
+    </Route>
+  )
+);
+
+export default router;
